requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.1
psycopg2-binary==2.9.9
gunicorn==20.1.0
python-dotenv==1.0.0 